refactor: simplify SEO canonical URL handling and update package dependencies
